Files for parsing and formatting DailyDialog data
I included the two raw .txt files needed to run the .py code, but I think those may already be in main branch. I also included an example .txt file with a formatting similar to one of those we discussed in class. There are further instructions for hopefully easy manipulation in the .py file!